requirements.txt
alembic==1.17.2
bidict==0.23.1
blinker==1.9.0
certifi==2025.11.12
click==8.3.1
dnspython==2.8.0
email-validator==2.3.0
eventlet==0.40.3
Flask==3.1.2
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.2.4
h11==0.16.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
pillow==12.0.0
python-dotenv==1.2.1
python-engineio==4.12.3
python-socketio==5.14.3
simple-websocket==1.1.0
SQLAlchemy==2.0.44
typing_extensions==4.15.0
Werkzeug==3.1.3
wsproto==1.3.2
WTForms==3.2.1

# Testing dependencies
pytest==8.0.0
pytest-flask==1.3.0
hypothesis==6.98.0
pytest-cov==4.1.0
